Fix shared TestLease and TestLeaseActor classes
