+++ src/contrib/cluster/Akka.Cluster.Sharding.Tests/ShardWithLeaseSpec.cs
@@ -13,7 +13,6 @@ using Akka.Actor;
 using Akka.Cluster.Tools.Singleton;
 using Akka.Configuration;
 using Akka.Coordination;
-using Akka.Coordination.Tests;
 using Akka.Event;
 using Akka.TestKit;
 using Akka.Util;
