+++ src/contrib/cluster/Akka.Cluster.Tools.Tests/Singleton/ClusterSingletonLeaseSpec.cs
@@ -15,7 +15,6 @@ using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.Cluster.Tools.Singleton;
 using Akka.Configuration;
-using Akka.Coordination.Tests;
 using Akka.Event;
 using Akka.TestKit;
 using Akka.TestKit.TestActors;
