+++ src/core/Akka.Cluster.Tests.MultiNode/SBR/LeaseMajority5NodeSpec.cs
@@ -12,7 +12,6 @@ using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.Cluster.TestKit;
 using Akka.Configuration;
-using Akka.Coordination.Tests;
 using Akka.MultiNode.TestAdapter;
 using Akka.Remote.TestKit;
 using Akka.Remote.Transport;
