+++ src/core/Akka.Docs.Tests/Utilities/LeaseActorDocSpec.cs
@@ -9,8 +9,8 @@ using System;
 using System.Threading.Tasks;
 using Akka.Actor;
 using Akka.Coordination;
-using Akka.Coordination.Tests;
 using Akka.Event;
+using Akka.TestKit;
 using Akka.TestKit.Xunit2;
 using FluentAssertions.Extensions;
 using Xunit;
